Add files via upload
This program receives an input from a user inform by presenting some options to choose from. After checking for the entered menu if available will ask the user to insert the money he/she has and if the money is enough, it will make coffee for the user, otherwise, it will make refunds.